requirements.txt
# Core Dependencies
openai>=1.0.0
langchain>=0.0.350
python-dotenv>=1.0.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Repository Management
GitPython>=3.1.0

# External Research API
tavily-python>=0.3.0

# Utility Libraries
requests>=2.28.0
pathlib2>=2.3.0
argparse>=1.4.0

# Text Processing
tiktoken>=0.5.0

# Optional: For better performance
numpy>=1.24.0
pandas>=2.0.0